requirements.txt
numpy >= 1.17.3
moviepy == 2.1.1
demucs == 4.0.1
psutil == 6.0.0
torch >= 2.0.0,< 2.5
pyannote.audio == 3.3.2
ctranslate2 >= 4.1.0, <= 4.4.0
faster-whisper == 1.1.1
transformers >= 4.40, <= 4.45.2

spacy[ja] == 3.7.6

iso639-lang == 2.3.0
edge-tts == 6.1.12
python-dotenv
openai